Trait oak_core::samplestore::SampleStore

pub trait SampleStore {
    // Required methods
    fn record(&mut self, value: u64);
    fn percentile(&self, n: f64) -> Option<u64>;
}

Required Methods

fn record(&mut self, value: u64)

Records a new data point value.

If the store has already been filled, it will probabilistically replace one of the existing entries based on reservoir sampling.

fn percentile(&self, n: f64) -> Option<u64>

Gets the n-th percentile.

n needs to be between 0 and 100, otherwise None is returned; also, the store must have at least one recorded data point.